+++ Fleece/Integration/MValue.hh
@@ -12,7 +12,6 @@
 
 #pragma once
 #include "fleece/Fleece.hh"
-#include "fleece/Base.h"
 #include "fleece/slice.hh"
 
 namespace fleece {
